src/backend/collector/requirements.txt
asyncio-mqtt>=0.16.2,<1.0
asyncpg>=0.28,<1.0
python-dotenv>=1.0
